조건문 다형성으로 쪼개기 #1
