+++ src/main/java/me/whiteship/refactoring/_03_long_function/_13_replace_conditional_with_polymorphism/MarkdownPrinter.java
@@ -1,0 +1,46 @@
+package me.whiteship.refactoring._03_long_function._13_replace_conditional_with_polymorphism;
+
+import java.io.FileWriter;
+import java.io.IOException;
+import java.io.PrintWriter;
+import java.util.List;
+
+public class MarkdownPrinter extends StudyPrinter {
+
+    public MarkdownPrinter(int totalNumberOfEvents, List<Participant> participants) {
+        super(totalNumberOfEvents, participants);
+    }
+
+    @Override
+    public void execute() throws IOException {
+        try (FileWriter fileWriter = new FileWriter("participants.md");
+             PrintWriter writer = new PrintWriter(fileWriter)) {
+
+            writer.print(header(this.participants.size()));
+
+            this.participants.forEach(p -> {
+                String markdownForHomework = getMarkdownForParticipant(p);
+                writer.print(markdownForHomework);
+            });
+        }
+    }
+
+    private String header(int totalNumberOfParticipants) {
+        StringBuilder header = new StringBuilder(String.format("| 참여자 (%d) |", totalNumberOfParticipants));
+
+        for (int index = 1; index <= this.totalNumberOfEvents; index++) {
+            header.append(String.format(" %d주차 |", index));
+        }
+        header.append(" 참석율 |\n");
+
+        header.append("| --- ".repeat(Math.max(0, this.totalNumberOfEvents + 2)));
+        header.append("|\n");
+
+        return header.toString();
+    }
+
+    private String getMarkdownForParticipant(Participant p) {
+        return String.format("| %s %s | %.2f%% |\n", p.username(), checkMark(p),
+                p.getRate(this.totalNumberOfEvents));
+    }
+}
