--- src/main/java/me/whiteship/refactoring/_03_long_function/_13_replace_conditional_with_polymorphism/PrinterMode.java
@@ -1,5 +1,0 @@
-package me.whiteship.refactoring._03_long_function._13_replace_conditional_with_polymorphism;
-
-public enum PrinterMode {
-    CONSOLE, CVS, MARKDOWN
-}
