+++ src/main/java/me/whiteship/refactoring/_03_long_function/_13_replace_conditional_with_polymorphism/StudyDashboard.java
@@ -29,7 +29,7 @@ public class StudyDashboard {
 
     private void print() throws IOException, InterruptedException {
         checkGithubIssues(getGhRepository());
-        new StudyPrinter(this.totalNumberOfEvents, this.participants, PrinterMode.MARKDOWN).execute();
+        new ConsolePrinter(this.totalNumberOfEvents, this.participants).execute();
     }
 
     private GHRepository getGhRepository() throws IOException {
